@@ -1,9 +1,6 @@
-module.exports = {
-  future: {
-    // removeDeprecatedGapUtilities: true,
-    // purgeLayersByDefault: true,
-  },
-  purge: ['./src/**/*.{js,jsx,ts,tsx}', './public/index.html'],
+/** @type {import('tailwindcss').Config} */
+export default {
+  content: ['./index.html', './src/**/*.{js,ts,jsx,tsx}'],
   darkMode: 'class',
   theme: {
     fontFamily: {
@@ -14,6 +11,5 @@ module.exports = {
     },
     extend: {},
   },
-  variants: {},
   plugins: [],
 }
