@@ -1,11 +1,10 @@
-import React from 'react'
 import FeelingButton from './FeelingButton'
 
-const FeelingInDesk: React.FC = () => {
+const FeelingInDesk = () => {
   return (
-    <div className="flex justify-center w-full mt-8 justify-items-center gap-x-4">
-      <FeelingButton text="Find with Google" />
-      <FeelingButton text="I feel lucky" />
+    <div className="flex justify-center pt-4">
+      <FeelingButton text="Google Search" />
+      <FeelingButton text="I'm Feeling Lucky" />
     </div>
   )
 }
